Hackathon_1/backend/pyproject.toml
[project]
name = "rag-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi>=0.127.0",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "ultimate-sitemap-parser>=1.6.0",
    "uvicorn>=0.40.0",
]
